ci: move e2e code to separate file
