+++ ci/e2e.go
@@ -1,0 +1,62 @@
+package main
+
+import (
+	"fmt"
+	"path"
+)
+
+func (m *Ci) Etoe(test Optional[string]) *Container {
+	image := m.Build().ContainerImage("").
+		WithExposedPort(10000).
+		WithMountedFile("/etc/openmeter/config.yaml", dag.Host().File(path.Join(root(), "e2e", "config.yaml"))).
+		WithServiceBinding("kafka", dag.Kafka(KafkaOpts{Version: kafkaVersion}).Service()).
+		WithServiceBinding("clickhouse", clickhouse())
+
+	api := image.
+		WithExposedPort(8080).
+		WithExec([]string{"openmeter", "--config", "/etc/openmeter/config.yaml"}).
+		AsService()
+
+	sinkWorker := image.
+		WithServiceBinding("redis", redis()).
+		WithServiceBinding("api", api). // Make sure api is up before starting sink worker
+		WithExec([]string{"openmeter-sink-worker", "--config", "/etc/openmeter/config.yaml"}).
+		AsService()
+
+	args := []string{"go", "test", "-v"}
+
+	if t, ok := test.Get(); ok {
+		args = append(args, "-run", fmt.Sprintf("Test%s", t))
+	}
+
+	args = append(args, "./e2e/...")
+
+	return dag.Go(GoOpts{
+		Container: dag.Go(GoOpts{Version: goVersion}).
+			WithSource(m.Source).
+			Container().
+			WithServiceBinding("api", api).
+			WithServiceBinding("sink-worker", sinkWorker).
+			WithEnvVariable("OPENMETER_ADDRESS", "http://api:8080"),
+	}).
+		Exec(args)
+}
+
+func clickhouse() *Service {
+	return dag.Container().
+		From(fmt.Sprintf("clickhouse/clickhouse-server:%s-alpine", clickhouseVersion)).
+		WithEnvVariable("CLICKHOUSE_USER", "default").
+		WithEnvVariable("CLICKHOUSE_PASSWORD", "default").
+		WithEnvVariable("CLICKHOUSE_DB", "openmeter").
+		WithExposedPort(9000).
+		WithExposedPort(9009).
+		WithExposedPort(8123).
+		AsService()
+}
+
+func redis() *Service {
+	return dag.Container().
+		From(fmt.Sprintf("redis:%s-alpine", redisVersion)).
+		WithExposedPort(6379).
+		AsService()
+}
